движение к абстрации, в связи со свойствами питона убираем ненужные классы
