+++ main.py
@@ -1,22 +1,28 @@
 # Main class test
 
-import logic
+import handler
 import GUI
 import sys
 
 
 class Main:
-    def start(self, url):
-        runner = logic.Runner(), GUI.GUI()
-        runner[0].parse(url)
-        runner[0].handler(runner[0].akk)
-        runner[1].output(runner[0].out)
+    def starter(self, url):
+        runner = handler.Runner()
+        gui = GUI.GUI()
+        runner.parse(url)
+        runner.handler(runner.agr)
+        gui.output()
+
 
 
 try:
     print('start')
     m = Main()
-    m.start('3.csv')
+    m.starter('3.csv')
+
+except Exception:
+    print(sys.exc_info())
+
 ##    a = sys.argv
 ##    print(a)
 ##    if (a is None):
@@ -26,6 +32,3 @@ try:
 ##        print('auto')
 ##        m.start(a[1])
 ##    print('end')
-except Exception:
-    print(sys.exc_info())
-
